backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "clip",
    "fastapi>=0.117.1",
    "numpy>=2.3.3",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "pydantic[email]>=2.11.9",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "uvicorn>=0.36.0",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
